BLIP_ViT_L_Finetuning/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.21.0
datasets>=2.0.0
Pillow>=8.3.0
wandb>=0.13.0
scikit-learn>=1.0.0
numpy>=1.21.0
pyyaml>=6.0
accelerate>=0.12.0
evaluate>=0.4.0 